+++ src/gui/win32/mod.rs
@@ -36,6 +36,7 @@ pub fn run_wizard(config: WizardConfig, installer: &mut Installer) -> Result<()>
             page,
             on_enter,
             on_before_leave,
+            skip_if,
         } = configured;
         let page = match page {
             WizardPage::Install {
@@ -55,6 +56,7 @@ pub fn run_wizard(config: WizardConfig, installer: &mut Installer) -> Result<()>
             page,
             on_enter,
             on_before_leave,
+            skip_if,
         });
     }